仲裁代码优化，取整去小数

+++ 51fbapi-biz/src/main/java/com/ald/fanbei/api/biz/service/impl/ArbitrationServiceImpl.java
@@ -313,7 +313,7 @@ public class ArbitrationServiceImpl extends BaseService implements
 		    result.put("amtPenalty", afBorrowCashDo.getOverdueAmount().multiply( BigDecimalUtil.ONE_HUNDRED).intValue());// 罚息
 	    } else {
 		    result.put("amtPenalty",new BigDecimal(0.015).multiply(new BigDecimal(afBorrowCashDo.getOverdueDay()))
-					.multiply(afBorrowCashDo.getAmount()).divide(new BigDecimal(360),3,RoundingMode.HALF_UP).multiply(BigDecimalUtil.ONE_HUNDRED));
+					.multiply(afBorrowCashDo.getAmount()).divide(new BigDecimal(360),3,RoundingMode.HALF_UP).multiply(BigDecimalUtil.ONE_HUNDRED).intValue());
 	    }
 	    result.put("amtService",  afBorrowCashDo.getAmount().subtract(afBorrowCashDo.getArrivalAmount()).multiply( BigDecimalUtil.ONE_HUNDRED).intValue());// 服务费
